Calibrate four VC5 functions in an isolated Docker toolchain
Run VC5 RTM 11.00.7022 and linker 5.00.7022 through a pinned wibo
container. Match the four complete code sections, including the original
projection padding, with verified address fixups and resolved span hashes.
Correct the source evaluation order and actual VC5 backend/flag assumptions.

Record two fresh passing VC5 builds, the private regression baseline,
35 passing tooling tests, source annotation lint, and a passing modern
x86 finite-behavior smoke. Keep compiler files and retail data private.
Whole-executable reconstruction and retail CRT startup matching remain open.

+++ engine/src/plane_coordinate.c
@@ -1,7 +1,8 @@
 #include "baseline.h"
 
 // FUNCTION: WMAIN 0x0040dcee
-/* Candidate, not yet verified with VC5. Valid axis is 0, 1 or 2.
+/* VC5-calibrated source. Valid axis is 0, 1 or 2.
+ * Double evaluation in the Y/Z divisions retains the observed x87 load order.
  * These separate stores and the switch are observable in the original codegen.
  * VC5's x87 equality sequence also takes the early return for unordered input;
  * a modern compiler's ordered C equality does not reproduce that behavior.
@@ -22,13 +23,13 @@ float op_plane_coordinate(int axis, float *vertex, float *normal, float *point)
     case 1:
         point[1] = (vertex[0] - point[0]) * normal[0];
         point[1] = (vertex[2] - point[2]) * normal[2] + point[1];
-        point[1] = point[1] / normal[1];
+        point[1] = (float)((double)point[1] / normal[1]);
         point[1] = point[1] + vertex[1];
         break;
     case 2:
         point[2] = (vertex[0] - point[0]) * normal[0];
         point[2] = (vertex[1] - point[1]) * normal[1] + point[2];
-        point[2] = point[2] / normal[2];
+        point[2] = (float)((double)point[2] / normal[2]);
         point[2] = point[2] + vertex[2];
         break;
     }
